[SCRUM-35] refact: Service와 Impl로 Refact For SOLID
