+++ src/main/java/example/demo/domain/chat/gemini/service/QnAService.java
@@ -1,53 +1,15 @@
 package example.demo.domain.chat.gemini.service;
 
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
-import org.springframework.web.reactive.function.client.WebClient;
-
-import java.util.Map;
 
 /**
- * Gemini API Service
+ * Gemini Service Interface
  *
  * @Author : 박재성
- * @Since : 02/15/2025
+ * @Since : 02/16/2025
  */
 
 @Service
-public class QnAService {
-    // Access to APIKey and URL [Gemini]
-    @Value("${gemini.api.url}")
-    private String geminiApiUrl;
-
-    @Value("${gemini.api.key}")
-    private String geminiApiKey;
-
-    private final WebClient webClient;
-
-    public QnAService(WebClient.Builder webClient) {
-        this.webClient = webClient.build();
-    }
-
-    public String getAnswer(String question) {
-        // Construct the request payload
-        Map<String, Object> requestBody = Map.of(
-                "contents", new Object[] {
-                        Map.of("parts", new Object[] {
-                                Map.of("text", question)
-                        })
-                }
-        );
-
-        // Make API Call
-        String response = webClient.post()
-                    .uri(geminiApiUrl + geminiApiKey)
-                    .header("Content-Type", "application/json")
-                    .bodyValue(requestBody)
-                    .retrieve()
-                    .bodyToMono(String.class)
-                    .block();
-
-        // Return response
-        return response;
-    }
+public interface QnAService {
+    String getAnswer(String question);
 }
